+++ yggdrasil_decision_forests/port/python/CHANGELOG.md
@@ -4,11 +4,22 @@
 
 ### Feature
 
+-   Expose `validate_hyperparameters()` on the learner.
+-   Clarify which parameters in the learner are optional.
 -   Add support in JAX FeatureEncoder for non-string categorical feature values.
+-   Improve performance of Isolation Forests.
+-   Models can be serialized/deserialized to/from bytes with `model.serialize()`
+    and `ydf.deserialize_model`.
+-   Models can be pickled safely.
+-   Native support for Xarray as a dataset format for all operations (e.g.,
+    training, evaluation, predictions).
+-   The output of `model.to_jax_function` can then be converted to a TensorFlow
+    Lite model.
 
 ### Fix
 
 -   Fix parsing of multidimensional ragged inputs.
+-   Fix isolation forest hyperparameter defaults.
 
 ## 0.6.0 - 2024-07-04